backend/generate-contract/requirements.txt
python-docx==1.1.2
psycopg2-binary==2.9.9
requests==2.31.0
